app/requirements.txt
catboost
shap
pandas
numpy
Flask
Flask-Markdown
Flask-Session
requests
pillow